+++ CHANGELOG.md
@@ -60,6 +60,7 @@ and this project adheres to
 - cosmwasm-std: Create `Fraction<T>` trait to represent a fraction `p`/`q` with
   integers `p` and `q`. `Decimal` now implements `Fraction<u128>`, which
   provides public getters `::nominator()` and `::denominator()`.
+- cosmwasm-std: Add `Decimal::inv` that returns `1/d` for decimal `d`.
 
 [#692]: https://github.com/CosmWasm/cosmwasm/issues/692
 [#706]: https://github.com/CosmWasm/cosmwasm/pull/706